Added 1296 amp embedded OLED wattmeter with SWR analog output signal for amp protection
Built this amp 11/2020 and installed a RF wattmeter assembly with OLED display on the front panel.  Uses the PSoC5 KitProg board for a compact solution.  Measuring 14V, 28V, ADL5519 detector temp and a DAC with buffer amp was configured to output an analog voltage corresponding to SWR range 1-5.  This is wired to the W6PQL Ultimate Amplifier Control Board Load Fail input to trip the amplifier in Hi SWR situation.  A current sensor and power reset relay yet to be added.