Tweaked the Current Zero Offset handling to enable ability to trim the value.  Before it ignored any UI value supplied.
The volt, temp and current cal and offset values are only stored with limited precision in EEPROM so range is X.XX. (Stored as a 4 byte ASCII value, not as a number)
Entering -0.008 for the No Load Current in the Desktop UI Config page bumps up a 0.4A current reading to 0.5A but it can only be stored as -0.01 or 0.00 so the value read after Save to Meter may change.  best to only use 2 decimal places max.  The minus sign I think is also lost.  This will need to be changed later for more precision and to support negative values. Probably store into EEPROM these as 16 bit signed integers multiplying the actual value by 100 to avoid the decimal point, then divide by 100 on read from EEPROM.  The actual values are float32 type.