Teensy 4.1 ported from latest PS0C5 code 12/1 (working!)
I ported the PSoC5 latest C code to the Teensy 4.1 Arduino 1.8.13 IDE.

The PSoC5 version used a modified version of the original Nextion Arduino library converted to C and includes only the most used parts.  The Nextion files are placed in the Documents\Arduino\libraries folder.

The SS1306 OLED files are also in C for the PSoC5 and do not not work on the Arduino as is.  The //#define SSD1306_OLED is commented out so it is bypassed for now and only used if you had a OLED connected to the I2C bus.

Happy to say that is is now working with some issues to tweak and fix still.  It will operate 100% it seems without any display using the Desktop App.  The Nextion screen is also working well except when viewing the graphing page which causes a display lockup and you need to reset the CPU.  Have not tested the calibration input pages yet, they call the same functions the Desktop does so should work as long as the comms work OK between the Nextion and CPU.

So testing on this Teensy 4.1 Arduino still ongoing but it looks like 95% or better is working good, including the EEPROM which is where rmost of the conversion work was.  Lots of Arduino IDE warnings to sift through and fix if warranted.