OTRSP now working (Mostly)
Still need to make the band decoder output IO pins work and convert the BCD AUX command input to BCD properly.  Today is processed as decimal value but it will change the Wattmeter band.  The CW and PTT from USB are not functional.   If the radio does not change band and you are running WSJTX with the Desktop App, the Desktop will send the meter the current radio band and thus set the Meter to match again.  This is the same as changing a the meter band from the touchscreen or the desktop app using the band button.  The meter always takes the last command from any source.