Resolved 99% of page sync issues
If you cycle through the Nextion pages really fast it is possible to miss an event.  I think the event is never sent out but the screen changes.  The best way to deal with this is to have the CPU make the actual change rather than the current way of the Nextion changing pages for sake iof instant response feel speed and sending out a notice of that change.  Placed strategic sendme command retry once on fail to reduce occurance of a page mismatch.  Long term will shift to a hidden API page with nothing but global variables and only address those variables from the CPU so always on the same page from the CPU perspective.