Added full featured Band Decoder functionality
Now either OTRSP or the 6 Band-Input pins can work to change bands and will send data out 3x 8-bit ports, A, B and C.

By default port A will mirror the input pins, useful for intercepting a BCD or 1-of-8 Band decoder from a radio on the input port then pass it on through Port A to a stack of transverters, or the Q5 Signal 5-Band transverter.

2 additional ports of 8 pins can serve amps and antennas or power meter coupler selection.

A cool feature is the variety of translation modes for every port.  Each port generally has 4 translation modes. 1. Transparent
2. 1-of-8 decode
3. Custom pattern (per band),
4. OTRSP value from AUX commands.

This is all stored in EEPROM.  Also fixed a bug where I calculated the EEPROM size of the main data table wrong causing a lack of EEPROM storage.  There is plenty after this fix.

This work is tailored for the Teensy 4.1 and changes wil be ported back to the PSoC5 platform.  Other Arduinos might work but I am using 2 USB serial ports (Main data an OTRSP), 1 hardware port (Nextion, optional) and over 1K of EEPROM.

The Toggle Serial feature is used to suppress the data ouput to make seeing debug messages easier.
 This is disabled for now becaue teh Desktop app cannot seems to rx chracters or it cannot serve commands including turning the data back on.  Teh CPU is OK, can manually toggle data OK.

Work remains to create configuration screens on the Nextion and the Desktop App. New serial port commands 62-69 exist to support this.

An area left to debug is the CW and PTT control from N1MM serial ports using DTR and RTS signals.  This works on the PSoC5 but I have yet to make it work on the Arduino.