Added remote commands and Desktop App Config Screen Suppprt for PTT Options
Could use some screen cleanup but what is there works.  Once exception is Toggle Serial Data button.  It has an issue on the Pythn side so will only send command to turn on data for now.  Can manually send 100,120,239,0 in monitor windows to turn off.  100,120,239,1 to turn it back on.