Added new code for Configurable IP address stored in EEPROM.
See .ino file header comments for more details.  Goal when complete is to not require hard coded IP addresses that require the CPU to be reprogrammed each time the desktop app IP address changes.  This could be when you swap computers, maybe want to run the desktop app on different machine.  Ideally it would just respond to the desktop app that queries it.  I may look at that in the near future.  I started with the static hard coded IPs for simplicity.  DHCP is possible for the decoder side.  The desktop needs to know the address so as long as you have a DHCP reservation or can otherwise find the decoder address (router table access) then that is viable also.  That will theoretically work today.